refactor: reorganize dependency injection modules and update application class structure
