+++ app/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
 
     <uses-permission android:name="android.permission.INTERNET" />
     <application
-        android:name=".MovioApp"
+        android:name=".di.MovioApp"
         android:screenOrientation="unspecified"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
