--- app/src/main/java/com/madrid/movio/data.kt
@@ -1,2 +1,0 @@
-package com.madrid.movio
-
