+++ app/src/main/java/com/madrid/movio/di/AppModule.kt
@@ -1,0 +1,14 @@
+package com.madrid.movio.di
+
+
+import com.madrid.data.roomModule
+import org.koin.dsl.module
+
+val AppModule = module {
+    includes(
+        dataModule,
+        domainModule,
+        presentationModule,
+        roomModule
+    )
+}
