+++ app/src/main/java/com/madrid/movio/di/DomainModule.kt
@@ -1,0 +1,26 @@
+package com.madrid.movio.di
+
+
+import com.madrid.domain.usecase.GetExploreMoreMovieUseCase
+import com.madrid.domain.usecase.GetRecommendedMovieUseCase
+import com.madrid.domain.usecase.mediaDeatailsUseCase.MovieDetailsUseCase
+import com.madrid.domain.usecase.searchUseCase.ArtistUseCase
+import com.madrid.domain.usecase.searchUseCase.MovieUseCase
+import com.madrid.domain.usecase.searchUseCase.PreferredMediaUseCase
+import com.madrid.domain.usecase.searchUseCase.RecentSearchUseCase
+import com.madrid.domain.usecase.searchUseCase.SeriesUseCase
+import com.madrid.domain.usecase.searchUseCase.TrendingMediaUseCase
+import org.koin.dsl.module
+
+
+val domainModule = module {
+    single { ArtistUseCase(get()) }
+    single { MovieUseCase(get()) }
+    single { SeriesUseCase(get()) }
+    single { PreferredMediaUseCase(get()) }
+    single { RecentSearchUseCase(get()) }
+    single { TrendingMediaUseCase(get()) }
+    single { GetExploreMoreMovieUseCase(get()) }
+    single { GetRecommendedMovieUseCase(get()) }
+    single { MovieDetailsUseCase(get()) }
+}
