+++ app/src/main/java/com/madrid/movio/di/MovioApp.kt
@@ -1,7 +1,6 @@
-package com.madrid.movio
+package com.madrid.movio.di
 
 import android.app.Application
-import com.madrid.data.roomModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
 
@@ -10,7 +9,7 @@ class MovioApp : Application() {
         super.onCreate()
         startKoin {
             androidContext(this@MovioApp)
-            modules(app, roomModule)
+            modules(AppModule)
         }
     }
-} 
+}
